set timeout value from long type to int type

+++ jestful-core/src/main/java/org/qfox/jestful/core/annotation/Timeout.java
@@ -24,7 +24,7 @@ public @interface Timeout {
 	 * 
 	 * @return
 	 */
-	long value();
+	int value();
 
 	/**
 	 * 超时时间单位, 默认为秒
